Move nft creation flows to appropriate files, deprecate file, remove duplication.
